+++ Sources/MapboxNavigation/ExitView.swift
@@ -23,6 +23,8 @@ public class ExitView: StylableView {
         }
     }
     
+    @objc public dynamic var highlightColor: UIColor?
+    
     var side: ExitSide = .right {
         didSet {
             populateExitImage()
@@ -147,7 +149,8 @@ public class ExitView: StylableView {
     static func criticalHash(side: ExitSide,
                              styleID: String?,
                              dataSource: DataSource,
-                             traitCollection: UITraitCollection) -> String {
+                             traitCollection: UITraitCollection,
+                             isHighlighted: Bool = false) -> String {
         var appearance = ExitView.appearance(for: UITraitCollection(userInterfaceIdiom: .phone))
         if traitCollection.userInterfaceIdiom == .carPlay {
             let traitCollection = UITraitCollection(traitsFrom: [
@@ -160,6 +163,7 @@ public class ExitView: StylableView {
         
         var criticalProperties: [AnyHashable?] = [
             side,
+            isHighlighted,
             dataSource.font.pointSize,
             appearance.backgroundColor,
             appearance.foregroundColor,